Removing circular import
