+++ src/mouse_tracking/cli/utils.py
@@ -9,10 +9,9 @@ from mouse_tracking import __version__
 from mouse_tracking.core.config.pose_utils import PoseUtilsConfig
 from mouse_tracking.matching.match_predictions import match_predictions
 from mouse_tracking.pose import render
-from mouse_tracking.pose.convert import downgrade_pose_file
 from mouse_tracking.utils import fecal_boli, static_objects
 from mouse_tracking.utils.clip_video import clip_video_auto, clip_video_manual
-from mouse_tracking.utils.writers import filter_large_poses
+from mouse_tracking.utils.writers import downgrade_pose_file, filter_large_poses
 
 app = typer.Typer()
 CONFIG = PoseUtilsConfig()
