+++ src/mouse_tracking/core/config/pose_utils.py
@@ -36,4 +36,4 @@ class PoseUtilsConfig(BaseSettings):
     MIN_JABS_KEYPOINTS: int = 3
 
     # Large animals are rarely larger than 100px in our OFA
-    OFA_MAX_EXPECTED_AREA_PX = 150 * 150
+    OFA_MAX_EXPECTED_AREA_PX: int = 150 * 150
